+++ tests/test_unitaire/test_auth.py
@@ -2,25 +2,8 @@ import pytest
 
 from flask import session
 from server import research_club_in_clubs_by_email
+from tests.fixtures import club_one, clubs 
 
-
-@pytest.fixture
-def club_one():
-    club_one = {'name': 'club_test1_name',
-                'email': 'club_test1@mail.com',
-                'points': '15'}
-    return club_one
-
-
-@pytest.fixture
-def clubs():
-    clubs = [{'name': 'club_test1_name',
-              'email': 'club_test1@mail.com',
-              'points': '15'},
-             {'name': 'club_test2_name',
-              'email': 'club_test2@mail.com',
-              'points': '8'}]
-    return clubs
 
 
 def _login(client, email):
